requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
google-generativeai
requests
python-dotenv